+++ packages/shared/components/src/Input/index.ts
@@ -1,0 +1,9 @@
+/* eslint-disable @stylistic/padding-line-between-statements, react-refresh/only-export-components */
+export {
+  InputLeftElement,
+  type InputLeftElementProps,
+  InputRightElement,
+  type InputRightElementProps,
+} from './input-element';
+export { InputGroup, type InputGroupProps } from './input-group';
+export { Input, type InputProps } from './input';
